Revert "java: Primitive class loading"
This reverts commit dd71e99b2c4ed6a05c0cf02b029b3351e2fdcf94.

It seems to make us try to free() a pointer that was not malloc'd.
Probably one of the primitive class raw pointers gets converted to a
shared_ptr and as that goes out of scope, we attempt to free statically
allocated memory.

Spotted breakage on both OS X and Linux.

Signed-off-by: Pekka Enberg <penberg@iki.fi>